Refactor run(): extract parse_args + cmd_create, switch to u8
Behavior-preserving cleanup folding four small wins into one diff:

- parse_args: argv prepending, clap parse, and error/help routing
  collapsed into one function returning Result<Cli, u8>. run()
  shrinks to a parse-then-dispatch shape ready for verbs 2-7.
- cmd_create: the create handler in its own function. Sets the
  pattern for future subcommand handlers without committing to a
  module structure yet (premature with one verb).
- e.use_stderr() lifted into a `to_stderr` binding; previously
  called twice in adjacent lines.
- Return type i32 -> u8. Matches the Unix exit-code range and
  removes the `code as u8` cast in main.rs. Tests compile
  unchanged (literal 0 is type-inferred to u8).

Tests stay green; no behavior change.

+++ src/main.rs
@@ -11,5 +11,5 @@ fn main() -> ExitCode {
     let mut stdout = io::stdout();
     let mut stderr = io::stderr();
     let code = tenant::run(&args, &accounts, &mut stdout, &mut stderr);
-    ExitCode::from(code as u8)
+    ExitCode::from(code)
 }
